edu-feedback-api/requirements.txt
flask
flask-cors
torch
transformers
pytorch-lightning
gunicorn
uvicorn
google-cloud-storage
flask
functions-framework
